chore: extract datetime-local helpers
